# openapi.yaml
openapi: 3.0.3
info:
  title: pandia API
  description: |
    HTTP API for rendering Markdown documents with embedded diagrams to PDF/HTML.

    **`/render`** operates on files mounted into the container (via `-v`).
    It reads a `.md` file from disk, writes output files (`.html`, `.pdf`) next to it,
    and returns the list of generated file paths.

    **`/preview`** is stateless: it accepts raw Markdown as the request body and
    returns self-contained HTML with all images inlined as SVG and math rendered
    as MathML. No files are written to disk. This is designed for editor integrations
    (e.g. the VS Code extension) that need fast, on-the-fly rendering.
  version: 1.6.0
  license:
    name: MIT

servers:
  - url: http://localhost:3300
    description: Default local server

paths:
  /health:
    get:
      summary: Health check
      description: Returns `ok` if the server is running and ready to accept requests.
      operationId: health
      responses:
        '200':
          description: Server is healthy
          content:
            text/plain:
              schema:
                type: string
                enum: [ok]

  /render:
    post:
      summary: Render a Markdown file on disk to PDF/HTML
      description: |
        Reads an input `.md` file from the container's `/data` directory,
        renders it to the requested format(s), and writes the output files
        next to the input file. Returns the list of generated file paths.

        The container must be started with a volume mount (`-v "$PWD:/data"`)
        so that input files are accessible and output files are written back.
      operationId: renderFile
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RenderRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RenderRequest'
      responses:
        '200':
          description: Rendering succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Rendering failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Render a Markdown file (query parameter variant)
      description: |
        Same as `POST /render`, but parameters are passed as query strings.
        Useful for quick testing in a browser.
      operationId: renderFileGet
      parameters:
        - name: file
          in: query
          required: true
          description: Path to the input `.md` file (relative to `/data`)
          schema:
            type: string
          example: example.md
        - name: to
          in: query
          required: false
          description: Output format(s), comma-separated
          schema:
            type: string
            default: html
            example: pdf,html
        - name: maxwidth
          in: query
          required: false
          description: Maximum content width for HTML output
          schema:
            type: string
            default: 60em
      responses:
        '200':
          description: Rendering succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Rendering failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /preview:
    post:
      summary: Render Markdown content to self-contained HTML
      description: |
        Accepts raw Markdown as the request body and returns fully self-contained
        HTML. All diagram images are inlined as SVG, math is rendered server-side
        as MathML. No files are written to disk.

        This endpoint is designed for live preview integrations (e.g. the VS Code
        extension). It is stateless — each request is independent.
      operationId: preview
      parameters:
        - name: maxwidth
          in: query
          required: false
          description: Maximum content width for the HTML output
          schema:
            type: string
            default: 60em
      requestBody:
        required: true
        description: Raw Markdown content
        content:
          text/plain:
            schema:
              type: string
            example: |
              # Hello

              ```graphviz
              digraph { A -> B; }
              ```
      responses:
        '200':
          description: Rendered HTML
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Empty or missing request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Rendering failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    RenderRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          description: Path to the input `.md` file (relative to `/data`)
          example: example.md
        to:
          type: string
          description: Output format(s), comma-separated (`html`, `pdf`, or `html,pdf`)
          default: html
          example: pdf,html
        maxwidth:
          type: string
          description: Maximum content width for HTML output
          default: 60em

    RenderResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        files:
          type: array
          items:
            type: string
          description: List of generated output file paths
          example: [example.html, example.pdf]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Missing \"file\" parameter"